Move common helper function to opengever.base.
