+++ opengever/disposition/ech0160/utils.py
@@ -1,6 +1,5 @@
-from zope.component import getUtility
-from zope.schema.interfaces import IVocabularyFactory
 from opengever.base.behaviors.classification import IClassification
+from opengever.base.vocabulary import voc_term_title
 import hashlib
 
 
@@ -13,22 +12,6 @@ def file_checksum(filename, chunksize=65536):
             h.update(chunk)
             chunk = f.read(chunksize)
         return u'MD5', h.hexdigest()
-
-
-def voc_term_title(field, value):
-    """Returns the vocabulary term title for the given field and value.
-    """
-    if value is None:
-        return None
-    factory = None
-    if field.vocabulary:
-        factory = field.vocabulary
-    elif field.vocabularyName:
-        factory = getUtility(IVocabularyFactory, name=field.vocabularyName)
-    if factory is not None:
-        voc = factory(None)
-        return voc.getTerm(value).title
-    return value
 
 
 def set_classification_attributes(binding, obj):
